[FIX] UI: revert input type hints for backwards compatibility.
The data table backport has changed several type hints which could lead to compile-
or fatal errors when plugins implemented custom inputs which extended from concrete
inputs of the UI framework. This commit reverts these changes back to the previous
state. There is still one breaking change in `Input::withRequired()` which gained
an additional parameter that could not be reverted.